do not escape song titles
The cue sheet format does not standardize quote escaping in strings. mpd
appears to truncate the string at the first quote after the opening
quote, regardless of a preceding backslash. mpv parses until the last
last quote but also does not parse escaped quotes. For example, the
title `TITLE "\"text\""` will be displayed as `\` in mpd and `\"text\"`
in mpv.